version 2.0.2: Update code dealing with extensions and system locale
importsav is now able to get system locale and set proper R repository
link for installation of haven. Also, it can check whether given file
name ends with file extension or not and will automatically add
appropriate extension to the end of given file name. And the code has
been modified to identify whether given input string was intended for
a subcommand or a file name.